split files and add labels
